Mehmet2/requirements.txt
backoff==2.2.1
click==8.1.7
loguru==0.7.2
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
tiktoken==0.3.3
